+++ PatientManager/Controllers/UserController.cs
@@ -1,19 +1,21 @@
 ﻿
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using PatientManager.Core.Application.Interfaces.Services;
 using PatientManager.Core.Application.ViewModels.User;
 using PatientManager.Core.Application.Helpers;
+using PatientManager.Middlewares;
 
 namespace PatientManager.Controllers
 {
     public class UserController : Controller
     {
         private readonly IUserService _userService;
+        private readonly ValidateUserSession _validator;
 
-        public UserController(IUserService userService)
+        public UserController(IUserService userService, ValidateUserSession validator)
         {
             _userService = userService;
+            _validator = validator;
         }
 
         // GET: UserController
@@ -56,11 +58,17 @@ namespace PatientManager.Controllers
         // GET: UserController
         public async Task<ActionResult> Index()
         {
+            if (!_validator.isAdmin())
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
             return View(await _userService.Get());
         }
 
         public ActionResult Register()
         {
+            if (!_validator.isAdmin())
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
             return View();
         }
 
@@ -68,6 +76,9 @@ namespace PatientManager.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Register(UserSaveViewModel vm)
         {
+            if (!_validator.isAdmin())
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
             try
             {
                 await _userService.Add(vm);
@@ -82,6 +93,9 @@ namespace PatientManager.Controllers
         // GET: UserController/Edit/5
         public async Task<ActionResult> Edit(int id)
         {
+            if (!_validator.isAdmin())
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
             return View(await _userService.GetById(id));
         }
 
@@ -90,6 +104,9 @@ namespace PatientManager.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(int id, UserUpdateViewModel vm)
         {
+            if (!_validator.isAdmin())
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
             try
             {
                 await _userService.Update(vm, id);
@@ -104,6 +121,9 @@ namespace PatientManager.Controllers
         // GET: UserController/Delete/5
         public async Task<ActionResult> Delete(int id)
         {
+            if (!_validator.isAdmin())
+                return RedirectToRoute(new { controller = "Home", action = "Index" });
+
             return View(await _userService.GetById(id));
         }
 
@@ -114,6 +134,9 @@ namespace PatientManager.Controllers
         {
             try
             {
+                if (!_validator.isAdmin())
+                    return RedirectToRoute(new { controller = "Home", action = "Index" });
+
                 await _userService.Delete(id);
                 return RedirectToAction(nameof(Index));
             }
